#50: Refactor by renaming to rd_process.py.
