+++ src/raindrop_todoist_syncer/main.py
@@ -7,7 +7,7 @@ from raindrop_todoist_syncer.env_manage import EnvironmentVariablesFileManager
 from raindrop_todoist_syncer.rd_credentials import RaindropCredentialsManager
 from raindrop_todoist_syncer.rd_process import RaindropsProcessor
 from raindrop_todoist_syncer.rd_token import RaindropAccessTokenRefresher
-from raindrop_todoist_syncer.raindrop import RaindropClient
+from raindrop_todoist_syncer.rd_client import RaindropClient
 
 from raindrop_todoist_syncer.td_task_create import TodoistTaskCreator
