+++ tests/integration/test_main.py
@@ -8,7 +8,7 @@ import pytest
 
 from raindrop_todoist_syncer.db_manage import DatabaseManager
 from raindrop_todoist_syncer.main import main
-from raindrop_todoist_syncer.raindrop import RaindropClient
+from raindrop_todoist_syncer.rd_client import RaindropClient
 from raindrop_todoist_syncer.rd_process import RaindropsProcessor
 from raindrop_todoist_syncer.td_task_create import TodoistTaskCreator
 
@@ -257,7 +257,7 @@ class TestMainValid:
         """
         rc = RaindropClient()
         with patch(
-            "raindrop_todoist_syncer.raindrop.RaindropClient.stale_token",
+            "raindrop_todoist_syncer.main.RaindropClient.stale_token",
             return_value=7,
         ):
             stale_token = rc.stale_token()
