+++ tests/unit/test_raindrop_client.py
@@ -6,7 +6,7 @@ from requests import HTTPError
 import requests
 import tenacity
 
-from raindrop_todoist_syncer.raindrop import RaindropClient
+from raindrop_todoist_syncer.rd_client import RaindropClient
 
 
 @pytest.fixture
@@ -36,7 +36,7 @@ class TestInit:
 class TestStaleToken:
     def test_200_response(self, mocker, rd_client):
         with patch(
-            "raindrop_todoist_syncer.raindrop.RaindropClient._core_api_call"
+            "raindrop_todoist_syncer.rd_client.RaindropClient._core_api_call"
         ) as mock_call:
             mock_response = Mock()
             mock_response.status_code = 200
@@ -46,7 +46,7 @@ class TestStaleToken:
 
     def test_401_response(self, rd_client):
         with patch(
-            "raindrop_todoist_syncer.raindrop.RaindropClient._core_api_call",
+            "raindrop_todoist_syncer.rd_client.RaindropClient._core_api_call",
             side_effect=requests.exceptions.HTTPError(response=Mock(status_code=401)),
         ):
             result = rd_client.stale_token()
@@ -54,7 +54,7 @@ class TestStaleToken:
 
     def test_404_response(self, rd_client):
         with patch(
-            "raindrop_todoist_syncer.raindrop.RaindropClient._core_api_call",
+            "raindrop_todoist_syncer.rd_client.RaindropClient._core_api_call",
             side_effect=requests.exceptions.HTTPError(response=Mock(status_code=404)),
         ):
             with pytest.raises(HTTPError):
